+++ esqueleto.cabal
@@ -41,11 +41,10 @@ library
     other-modules:
         Database.Esqueleto.Internal.PersistentImport
         Database.Esqueleto.PostgreSQL.JSON.Instances
-        Database.Esqueleto.Experimental.CommonTableExpression
-        Database.Esqueleto.Experimental.Internal
-        Database.Esqueleto.Experimental.Join
-        Database.Esqueleto.Experimental.SqlSetOperation
-        Database.Esqueleto.Experimental.SubQuery
+        Database.Esqueleto.Experimental.From
+        Database.Esqueleto.Experimental.From.CommonTableExpression
+        Database.Esqueleto.Experimental.From.Join
+        Database.Esqueleto.Experimental.From.SqlSetOperation
         Database.Esqueleto.Experimental.ToAlias
         Database.Esqueleto.Experimental.ToAliasReference
         Database.Esqueleto.Experimental.ToMaybe
